Backtestify/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
APScheduler==3.10.4
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.1.3
bidict==0.23.1
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.7
dhanhq==1.3
dnspython==2.6.1
docker==7.1.0
eventlet==0.36.1
Flask==3.0.3
Flask-APScheduler==1.13.1
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-DotEnv==0.1.2
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
numpy==1.26.4
pandas==2.2.2
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-engineio==4.9.0
python-socketio==5.11.2
pytz==2024.1
redis==5.0.3
requests==2.31.0
setuptools==69.5.1
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==2.0.29
talipp==2.2.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
websockets==12.0
Werkzeug==3.0.2
wsproto==1.2.0
yarl==1.9.4
zope.event==5.0
zope.interface==6.4
